backend/requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Validation & Serialization
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
click==8.1.7
requests==2.31.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
ruff==0.1.9
black==23.12.1
